requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
distro==1.9.0
fastapi==0.128.3
fastembed==0.7.4
firebase-admin==6.7.0
filelock==3.20.3
flatbuffers==25.12.19
fsspec==2026.2.0
google-genai==1.14.0
grpcio==1.78.0
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.4.1
hyperframe==6.1.0
idna==3.11
jiter==0.13.0
loguru==0.7.3
mmh3==5.2.0
mpmath==1.3.0
numpy==2.4.2
onnxruntime==1.24.1
openai==2.17.0
packaging==26.0
pillow==11.3.0
portalocker==3.2.0
protobuf==6.33.5
py_rust_stemmers==0.1.5
pydantic==2.12.5
pydantic_core==2.41.5
PyYAML==6.0.3
qdrant-client==1.16.2
requests==2.32.5
shellingham==1.5.4
sniffio==1.3.1
starlette==0.52.1
sympy==1.14.0
tokenizers==0.22.2
tqdm==4.67.3
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
